Move event tests into separate file, so that I can re-use setup/destroy logic for event tests only.
