+++ Assets/MixedRealityToolkit.Tests/PlayModeTests/InteractableTests.cs
@@ -31,174 +31,17 @@ namespace Microsoft.MixedReality.Toolkit.Tests
         const string defaultInteractablePrefabAssetPath = "Assets/MixedRealityToolkit.Examples/Demos/UX/Interactables/Prefabs/Model_PushButton.prefab";
         const string radialSetPrefabAssetPath = "Assets/MixedRealityToolkit.SDK/Features/UX/Interactable/Prefabs/RadialSet.prefab";
 
-        GameObject cube;
-        Interactable interactable;
-        TestHand testHand;
-
         [SetUp]
         public void Setup()
         {
             PlayModeTestUtilities.Setup();
             TestUtilities.PlayspaceToOriginLookingForward();
-
-            cube = GameObject.CreatePrimitive(PrimitiveType.Cube);
-            cube.transform.position = Vector3.forward;
-
-            interactable = cube.AddComponent<Interactable>();
-            Assert.NotNull(interactable, "Failed to add interactable to cube");
-            testHand = new TestHand(Handedness.Right);
         }
 
         [TearDown]
         public void TearDown()
         {
-            GameObject.Destroy(cube);
             PlayModeTestUtilities.TearDown();
-        }
-
-        /// <summary>
-        /// Tests that an Interactable component can be added to a GameObject
-        /// at runtime, and an OnClick event handler can be added
-        /// </summary>
-        /// <returns></returns>
-        [UnityTest]
-        public IEnumerator TestAddOnClickRuntime()
-        {
-            PlayModeTestUtilities.PushHandSimulationProfile();
-            PlayModeTestUtilities.SetHandSimulationMode(HandSimulationMode.Gestures);
-
-            // This should not throw an exception
-            var interactable = cube.AddComponent<Interactable>();
-
-            // Subscribe to interactable's on click so we know the click went through
-            bool wasClicked = false;
-            interactable.OnClick.AddListener(() => { wasClicked = true; });
-
-            yield return testHand.Show(Vector3.zero);
-
-            CameraCache.Main.transform.LookAt(interactable.transform);
-            yield return testHand.SetGesture(ArticulatedHandPose.GestureId.Pinch);
-            yield return testHand.SetGesture(ArticulatedHandPose.GestureId.Open);
-
-            Assert.True(wasClicked);
-
-            PlayModeTestUtilities.PopHandSimulationProfile();
-        }
-
-        [UnityTest]
-        public IEnumerator TestGrabEvents()
-        {
-            var grabReceiver = interactable.AddReceiver<InteractableOnGrabReceiver>();
-            interactable.gameObject.AddComponent<NearInteractionGrabbable>();
-            bool didGrab = false;
-            grabReceiver.OnGrab.AddListener(() => didGrab = true);
-            bool didRelease = false;
-            grabReceiver.OnRelease.AddListener(() => didRelease = true);
-
-            yield return testHand.Show(interactable.transform.position);
-            yield return testHand.SetGesture(ArticulatedHandPose.GestureId.Pinch);
-            yield return testHand.SetGesture(ArticulatedHandPose.GestureId.Open);
-            yield return testHand.Hide();
-            Assert.True(didGrab, "Did not receive grab event");
-            Assert.True(didRelease, "Did not receive release event");
-            GameObject.Destroy(cube);
-            yield return null;
-        }
-
-        public IEnumerator TestHoldEvents()
-        {
-            // Hold
-            var holdReceiver = interactable.AddReceiver<InteractableOnHoldReceiver>();
-            bool didHold = false;
-            holdReceiver.OnHold.AddListener(() => didHold = true);
-            yield return testHand.Show(interactable.transform.position);
-            yield return testHand.SetGesture(ArticulatedHandPose.GestureId.Pinch);
-            yield return new WaitForSeconds(holdReceiver.HoldTime);
-            yield return testHand.Hide();
-            Assert.True(didHold, "Did not receive hold event");
-            GameObject.Destroy(cube);
-            yield return null;
-        }
-
-        [UnityTest]
-        public IEnumerator TestPressEvents()
-        {
-            // Press
-            var pressReceiver = interactable.AddReceiver<InteractableOnPressReceiver>();
-            bool didPress = false;
-            pressReceiver.OnPress.AddListener(() => didPress = true);
-            yield return testHand.Show(interactable.transform.position);
-            yield return testHand.SetGesture(ArticulatedHandPose.GestureId.Pinch);
-            yield return testHand.Hide();
-            Assert.True(didPress, "did not receive press event");
-        }
-
-        [UnityTest]
-        public IEnumerator TestToggleEvents()
-        {
-            // Toggle
-            var toggleReceiver = interactable.AddReceiver<InteractableOnToggleReceiver>();
-            interactable.transform.position = Vector3.forward * 2f;
-            interactable.Dimensions = 2;
-            interactable.CanSelect = true;
-            interactable.CanDeselect = true;
-            bool didSelect = false;
-            bool didUnselect = false;
-            toggleReceiver.OnSelect.AddListener(() => didSelect = true);
-            toggleReceiver.OnDeselect.AddListener(() => didUnselect = true);
-            yield return testHand.Show(Vector3.forward);
-            yield return testHand.SetGesture(ArticulatedHandPose.GestureId.Open);
-            yield return null;
-            yield return testHand.SetGesture(ArticulatedHandPose.GestureId.Pinch);
-            yield return testHand.SetGesture(ArticulatedHandPose.GestureId.Open);
-            yield return testHand.SetGesture(ArticulatedHandPose.GestureId.Pinch);
-            yield return testHand.SetGesture(ArticulatedHandPose.GestureId.Open);
-            yield return testHand.Hide();
-            Assert.True(didSelect, "Toggle select did not fire");
-            Assert.True(didUnselect, "Toggle unselect did not fire");
-        }
-
-        [UnityTest]
-        public IEnumerator TestTouchEvents()
-        {
-            interactable.gameObject.AddComponent<NearInteractionTouchableVolume>();
-            var touchReceiver = interactable.AddReceiver<InteractableOnTouchReceiver>();
-            bool didTouch = false;
-            bool didUntouch = false;
-            touchReceiver.OnTouchStart.AddListener(() => didTouch = true);
-            touchReceiver.OnTouchEnd.AddListener(() => didUntouch = true);
-            yield return testHand.Show(Vector3.forward);
-            yield return testHand.SetGesture(ArticulatedHandPose.GestureId.Open);
-            yield return testHand.MoveTo(interactable.transform.position);
-            yield return testHand.MoveTo(Vector3.forward);
-            yield return testHand.Hide();
-            Assert.True(didTouch, "Did not receive touch event");
-            Assert.True(didUntouch, "Did not receive touch end event");
-        }
-
-        /// <summary>
-        /// Tests that an interactable component can be added to a GameObject
-        /// at runtime, and all receivers that extend ReceiverBase
-        /// and have event handlers can easily be added, and work
-        /// </summary>
-        /// <returns></returns>
-        [UnityTest]
-        public IEnumerator TestFocusEvents()
-        {
-            var onFocusReceiver = interactable.AddReceiver<InteractableOnFocusReceiver>();
-
-            bool didHover = false;
-            bool didUnHover = false;
-            onFocusReceiver.OnFocusOn.AddListener(() => didHover = true);
-            onFocusReceiver.OnFocusOff.AddListener(() => didUnHover = true);
-            CameraCache.Main.transform.LookAt(interactable.transform);
-            yield return null;
-            CameraCache.Main.transform.LookAt(Vector3.forward);
-            yield return null;
-            Assert.True(didHover, "Interactable did not receive hover event");
-            Assert.True(didUnHover, "Interactable did not receive un-hover event");
-
-            yield return null;
         }
 
         /// <summary>
